tests/wpt/web-platform-tests/html/syntax/charset/inheritance-bogus-meta-utf-8.html

<!doctype html>
<meta charset="utf-8" />
<title>Inheriting from UTF-8 parent</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
[
  {
    "title": "Child with bogus <meta charset>",
    "url": "resources/bogus-charset.html",
    "expected": "\uFFFD\n" // 0x00A2 in windows-1252 is 0xFFFD in UTF-8
  },
  {
    "title": "Child with bogus Content-Type charset",
    "url": "resources/bogus-charset-http.py",
    "expected": "\uFFFD\n"
  },
  {
    "title": "Child with bogus Content-Type charset, but valid <meta charset>",
    "url": "resources/bogus-charset-http-valid-meta.py",
    "expected": "\u045E\n"
  }
].forEach(({ title, url, expected }) => {
  async_test(t => {
    const frame = document.createElement("iframe");
    t.add_cleanup(() => frame.remove());
    frame.src = url;
    frame.onload = t.step_func_done(() => {
      assert_equals(frame.contentDocument.body.textContent, expected);
    });
    document.body.append(frame);
  }, title);
});
</script>
