LayoutTests/imported/w3c/web-platform-tests/html/syntax/charset/inheritance-bogus-meta-utf-8-expected.txt


PASS Child with bogus <meta charset>
FAIL Child with bogus Content-Type charset assert_equals: expected "\ufffd\n" but got "¢\n"
PASS Child with bogus Content-Type charset, but valid <meta charset>
